# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Support Tickets API",
        "description": "API para gerenciamento de chamados de suporte",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Listar todos os chamados",
                "operationId": "23abb903edea59ceb366f9f029c26991",
                "responses": {
                    "200": {
                        "description": "Lista de chamados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Criar novo chamado",
                "operationId": "0f82fdde1437d948b90892e1b714a1cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "priority",
                                    "requester_email"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Servidor fora"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "O servidor caiu"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ],
                                        "example": "high"
                                    },
                                    "requester_email": {
                                        "type": "string",
                                        "example": "user@test.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chamado criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Buscar chamado por ID",
                "operationId": "7231698c431b84bf9e4919256db5db1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chamado encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chamado não encontrado"
                    }
                }
            }
        },
        "/tickets/{id}/status": {
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Atualizar status do chamado",
                "operationId": "6421a5d4cbb8b7f49702b179f4154d86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "in_progress",
                                            "resolved",
                                            "closed"
                                        ],
                                        "example": "resolved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chamado não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Ticket": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "uuid-aqui"
                    },
                    "title": {
                        "type": "string",
                        "example": "Servidor fora"
                    },
                    "description": {
                        "type": "string",
                        "example": "O servidor caiu"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "in_progress",
                            "resolved",
                            "closed"
                        ]
                    },
                    "requesterEmail": {
                        "type": "string",
                        "example": "user@test.com"
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2026-05-04 19:40:39"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tickets",
            "description": "Tickets"
        }
    ]
}